requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
httpx>=0.26.0
langchain-core>=0.3.0
langchain-openai>=0.2.14
pyyaml>=6.0.1
jsonschema>=4.21.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
jsonpath-ng>=1.6.0
python-multipart>=0.0.9
pytest>=7.4.0
pytest-asyncio>=0.23.0
